+++ personal_graph/database/database_store/sql/delete-edge.sql
