+++ personal_graph/database/database_store/sql/insert-node.sql
