+++ personal_graph/database/database_store/sql/schema.sql
@@ -24,15 +24,3 @@ CREATE TABLE IF NOT EXISTS edges (
 
 CREATE INDEX IF NOT EXISTS source_idx ON edges(source);
 CREATE INDEX IF NOT EXISTS target_idx ON edges(target);
-
-begin;
-
-CREATE VIRTUAL TABLE IF NOT EXISTS nodes_embedding USING vss0(
-  vector_nodes(384)
-);
-
-CREATE VIRTUAL TABLE IF NOT EXISTS relationship_embedding USING vss0(
-  vector_relations(384)
-);
-
-commit;
