+++ personal_graph/database/database_store/sql/search-edges.sql
