+++ personal_graph/database/database_store/sql/search-node.template
