+++ personal_graph/database/database_store/sql/traverse.template
