--- personal_graph/database/sql/delete-edge-embedding.sql
@@ -1,1 +1,0 @@
-DELETE FROM relationship_embedding WHERE rowid = (?)
