--- personal_graph/database/sql/delete-node-embedding.sql
@@ -1,1 +1,0 @@
-DELETE FROM nodes_embedding where rowid = (?)
