--- personal_graph/database/sql/existing-edge.sql
@@ -1,1 +1,0 @@
-SELECT embed_id FROM edges WHERE source = ? AND target = ? AND label = ? AND attributes = json(?)
