--- personal_graph/database/sql/existing-node.sql
@@ -1,1 +1,0 @@
-SELECT embed_id FROM nodes WHERE id = ?
