--- personal_graph/database/sql/insert-edge-embedding.sql
@@ -1,1 +1,0 @@
-INSERT INTO relationship_embedding(rowid, vector_relations) VALUES(?, ?)
