--- personal_graph/database/sql/insert-node-embedding.sql
@@ -1,1 +1,0 @@
-INSERT INTO nodes_embedding(rowid, vector_nodes) VALUES (?,?);
